# pops-project-manager-api/contrato/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "POPS Project Manager API",
        "description": "API para gerenciamento de projetos e skills",
        "contact": {
            "name": "POPS Team",
            "email": "gyulia.piqueira@sptech.school"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8082/api",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/skills/enable/{id}": {
            "put": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "enableSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/skills/disable/{id}": {
            "put": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "disableSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/skills": {
            "get": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "listSkills",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "createSkill",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SkillCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Skill"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/skills/{id}": {
            "get": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "getSkillById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SkillResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "deleteSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "updateSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SkillUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/skills/type/{typeId}": {
            "get": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "listSkillsByType",
                "parameters": [
                    {
                        "name": "typeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/skills/pageable": {
            "get": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "listSkillsPagination",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SkillCreateRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "typeId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "name",
                    "typeId"
                ]
            },
            "Skill": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/SkillTypeEntity"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name",
                    "type"
                ]
            },
            "SkillTypeEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name"
                ]
            },
...
            "SkillUpdateRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "typeId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "SkillResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/SkillTypeResponse"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name",
                    "type"
                ]
            },
            "SkillTypeResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name"
                ]
            },
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    }
}